refactor: 5단계 - 자동차 경주(리팩토링)
1. 핵심 비지니 로직을 담당하는 RacingCar, Track 클래스를 domain 패키지로 이동
2. UI 로직을 담당하는 InputView, ResultView 클래스를 view 패키지로 이동
