+++ src/main/java/domain/Track.java
@@ -1,4 +1,4 @@
-package racingcar;
+package domain;
 
 import java.util.Arrays;
 import java.util.NoSuchElementException;
