+++ src/main/java/racingcar/Controller.java
@@ -1,5 +1,9 @@
 package racingcar;
 
+import domain.Track;
+import view.InputView;
+import view.ResultView;
+
 public class Controller {
 
     public static void main(String[] args) {
