+++ src/test/java/racingcar/RacingCarTest.java
@@ -1,5 +1,6 @@
 package racingcar;
 
+import domain.RacingCar;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
